Amishi Restaurant (first draft)
This web app was developed using Google AppScript  in tandem with Google Sheets to provide a POS food purchase service for my school's Amishi Mishi restaurant. The vendor can view details of the purchase, inventory or stock, payment methods and notes from the client. This is the first version and I look to update it soon.